+++ UVtools.Core/FileFormats/CXDLPFile.cs
@@ -17,7 +17,6 @@ using System.Drawing;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using UVtools.Core.Converters;