+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 30/06/2023 - v3.15.1
+
+- **File formats:**
+  - (Fix) CTB, CXDLPv4 and GOO: Unable to set some of the virtual bottom properties to 0
+  - (Fix) Anycubic and GOO: When open file in partial mode, the layer data was not set from the layer table
+  - (Fix) CXDLP: "Invalid printer model" error
+- (Add) File - Reset layers properties: Reset layers properties with the global properties of the file
+
+
 ## 26/06/2023 - v3.15.0
 
 - **File formats:**